CHG switching to the cpp library as default
